refactor(types): label
