+++ src/lib/builders/label/types.ts
@@ -1,0 +1,3 @@
+import type { createLabel } from './create';
+
+export type CreateLabelReturn = ReturnType<typeof createLabel>;
